+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Helper class: DirectoryHelper.
+- Helpers classes: DirectoryHelper, FileHelper.
 
 ## [1.0.0] - 2018-08-04
 ### Added